requirements.txt
requests
python-dotenv
langchain
pydantic
langchain-google-genai
langchain-anthropic
rich
scikit-learn
numpy
